Update homework_2.sql
TITLE: UofT-DSI | SQL - Homework 2

## What changes are you trying to make? (e.g. Adding or removing code, refactoring existing code, adding reports)

I have included answers to homework 2 in SQL code.

## What did you learn from the changes you have made?

I learned about the syntax of SQL, and the order the commands need to be in for the code to run. I learned how to create a new column, use a CASE statement,  filter columns by defining criteria and how to use INNER JOIN to join two tables through the common elements of a given column. I also learned that one has to look at everything carefully.

## Was there another approach you were thinking about making? If so, what approach(es) were you thinking of?

I tried using IN to search for 'pepper' in the product_name column. This didn't work! Instead we need product_name LIKE '%pepper%'.

## Were there any challenges? If so, what issue(s) did you face? How did you overcome it?

I had to look through the notes from class carefully to correct my syntax, and understand how the commands in SQL run.

## How were these changes tested?

I tested the changes in DB browser, and tried to look at the tables carefully.

## A reference to a related issue in your repository (if applicable)

## Checklist
- [ X] I can confirm that my changes are working as intended